==> code/01.cohortCancerComposition_piePlot.Rmd <==
###############################################################################################
Aim: Cancer composition of pan-cancer cohorts
Description: To make pie plots showing cancer composition of pan-cancer cohorts (Fig. 1a; Extended Data Fig. 1b).
###############################################################################################


# load required package
```{r}
library(data.table)
library(ggplot2)
library(readxl)

library(viridis)

```

# Set parameters and directories
```{r}
input_dir = "../02.Input/"
result_dir = "../03.Results/"
cutoff = 0.5

```

# load data 
```{r}
selected_cols = c("CancerType") # "SAMPLE_ID",
Chowell_df1 = read_excel(paste0(input_dir,'AllData.xlsx'), "Chowell_train")
Chowell_df1 = Chowell_df1[selected_cols]
Chowell_df2 = read_excel(paste0(input_dir,'AllData.xlsx'), "Chowell_test")
Chowell_df2 = Chowell_df2[selected_cols]
Chowell_df = rbind(Chowell_df1,Chowell_df2)

MSK_df1 = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK1")
MSK_df1 = MSK_df1[selected_cols]
MSK_df2 = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK12")
MSK_df2 = MSK_df2[selected_cols]
MSK_df = rbind(MSK_df1,MSK_df2)

Kato_df = read_excel(paste0(input_dir,'AllData.xlsx'), "Kato_panCancer")
Kato_df = Kato_df[selected_cols]

Pradat_df = read_excel(paste0(input_dir,'AllData.xlsx'), "Pradat_panCancer")
Pradat_df = Pradat_df[selected_cols]


MSK_nonICB_df = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK_nonICB")
MSK_nonICB_df = MSK_nonICB_df[selected_cols]

CancerTypes = unique(c(Chowell_df$CancerType, MSK_df$CancerType, Kato_df$CancerType, Pradat_df$CancerType))

```

# pie plot of cancer composition
```{r}
Chowell_cancerComposition <- table(factor(Chowell_df$CancerType, levels = CancerTypes))
Chowell_cancerComposition = as.data.frame(Chowell_cancerComposition)
colnames(Chowell_cancerComposition) = c('CancerType', 'Ratio')

MSK_cancerComposition <- table(factor(MSK_df$CancerType, levels = CancerTypes))
MSK_cancerComposition = as.data.frame(MSK_cancerComposition)
colnames(MSK_cancerComposition) = c('CancerType', 'Ratio')

Kato_cancerComposition <- table(factor(Kato_df$CancerType, levels = CancerTypes))
Kato_cancerComposition = as.data.frame(Kato_cancerComposition)
colnames(Kato_cancerComposition) = c('CancerType', 'Ratio')

Pradat_cancerComposition <- table(factor(Pradat_df$CancerType, levels = CancerTypes))
Pradat_cancerComposition = as.data.frame(Pradat_cancerComposition)
colnames(Pradat_cancerComposition) = c('CancerType', 'Ratio')

MSK_nonICB_cancerComposition <- table(factor(MSK_nonICB_df$CancerType, levels = CancerTypes))
MSK_nonICB_cancerComposition = as.data.frame(MSK_nonICB_cancerComposition)
colnames(MSK_nonICB_cancerComposition) = c('CancerType', 'Ratio')

####### pie chart colormap
pie_chart <- ggplot(MSK_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  geom_bar(stat="identity", width=1) +
  scale_fill_manual(values = rev(my.cols))+
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_colormap.pdf")), width = 1.8*3, height = 1.8*3) # width = 2.3, height = 2.7



####### pie chart of Chowell et al.
pie_chart <- ggplot(Chowell_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  scale_fill_manual(values = rev(my.cols))+
  geom_bar(stat="identity", width=1) +
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(legend.position = "none") +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_Chowell.pdf")), width = 1.8*1, height = 1.8*1) # width = 2.3, height = 2.7


####### pie chart of MSK1&2
pie_chart <- ggplot(MSK_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  geom_bar(stat="identity", width=1) +
  scale_fill_manual(values = rev(my.cols))+
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(legend.position = "none") +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_MSK.pdf")), width = 1.8*1, height = 1.8*1) # width = 2.3, height = 2.7


####### pie chart of Kato
pie_chart <- ggplot(Kato_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  scale_fill_manual(values = rev(my.cols))+
  geom_bar(stat="identity", width=1) +
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(legend.position = "none") +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_Kato.pdf")), width = 1.8*1, height = 1.8*1) # width = 2.3, height = 2.7


####### pie chart of Pradat
pie_chart <- ggplot(Pradat_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  scale_fill_manual(values = rev(my.cols))+
  geom_bar(stat="identity", width=1) +
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(legend.position = "none") +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_Pradat.pdf")), width = 1.8*1, height = 1.8*1) # width = 2.3, height = 2.7


####### pie chart of MSK non-ICB
pie_chart <- ggplot(MSK_nonICB_cancerComposition, aes(x="", y=Ratio, fill=CancerType)) +
  scale_fill_manual(values = rev(my.cols))+
  geom_bar(stat="identity", width=1) +
  coord_polar("y", start=0)
# Remove x and y axis labels and ticks
pie_chart <- pie_chart +
  theme_bw()+
  theme(legend.position = "none") +
  theme(axis.text.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks=element_blank(),
        panel.grid=element_blank(),
        panel.border = element_blank()) +
  labs(x = "", y = "")
# Save figure
ggsave(paste0(result_dir,paste0("cancerType_piePlot_MSK_nonICB.pdf")), width = 1.8*1, height = 1.8*1) # width = 2.3, height = 2.7


```
