==> code/07_2.PanCancer_LORIS_TMB_survivalAnalysis_ICB.Rmd <==
###############################################################################################
Aim: Pan-cancer survival analysis in ICB test cohort
Description: To study LORIS predictive power in merged Chowell et al. & MSK1 cohorts (Figs. 3, 4c,d; Extended Data Figs. 3, 4a-d, 7c-d, 8c,d)
(1) K-M survival curves for LORIS-high/TMB-high vs. LORIS-high/TMB-low vs. LORIS-low/TMB-high vs. LORIS-low/TMB-low groups
(2) Multi-variable hazard ratio forest plot in individual cancers
(3) 0.5-,1-,2-,3-,4-,5-year PFS/OS in individual cancers
(4) K-M survival curves for 0-10%,10-20%,20-50%,50-80%,80-90%,90-100%-quantile groups of LORIS
These analyses have been done in both pan-cancer patients and non-NSCLC patients.
###############################################################################################


# load required package
```{r}
library(data.table)
library(tidyverse)
library(survminer)
library(survival)
library(ggridges)
library(qvalue)
library(readxl)
library(verification)
library(pROC)
library(forestplot)
library(viridis)
```

# Set parameters and directories
```{r}
input_dir = "../02.Input/"
result_dir = "../03.Results/"

LLRmodelNA = 'LLR6'
absoluteCutoff = 0.5

```

# load data 
```{r}

cancer_type = 'nonNSCLC' # 'all'   'nonNSCLC'

test1_LLRscore_df = read_excel(paste0(result_dir,'PanCancer_',cancer_type,'_', LLRmodelNA,'_Scaler(StandardScaler)_prediction.xlsx'), "1", col_names = TRUE)
test2_LLRscore_df = read_excel(paste0(result_dir,'PanCancer_',cancer_type,'_', LLRmodelNA,'_Scaler(StandardScaler)_prediction.xlsx'), "2")
test1_RF6score_df = read_excel(paste0(result_dir,'PanCancer_',cancer_type,'_', 'RF6_Scaler(None)_prediction.xlsx'), "1")
test2_RF6score_df = read_excel(paste0(result_dir,'PanCancer_',cancer_type,'_', 'RF6_Scaler(None)_prediction.xlsx'), "2")
LLRscore_df = rbind(test1_LLRscore_df,test2_LLRscore_df)
RF6score_df = rbind(test1_RF6score_df,test2_RF6score_df)

selected_cols = c("SAMPLE_ID","Year of ICI treatment start", "TMB", "Systemic_therapy_history", "CancerType", "Age", "NLR", "Albumin", "Drug", "Sex", "PFS_Event", "PFS_Months", "OS_Event", "OS_Months")
test1_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "Chowell_test")
test1_info_df = test1_info_df[selected_cols]
test2_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK1")
test2_info_df = test2_info_df[selected_cols]
if (cancer_type == 'nonNSCLC'){
  test1_info_df = test1_info_df[test1_info_df['CancerType']!='NSCLC',]
  test2_info_df = test2_info_df[test2_info_df['CancerType']!='NSCLC',]
}

renamed_cols = c("SAMPLE_ID","Year_IO", "TMB", "Systemic_therapy_history", "CancerType", "Age", "NLR", "Albumin", "Drug", "Sex", "PFS_Event", "PFS_Months", "OS_Event", "OS_Months")
info_df = rbind(test1_info_df,test2_info_df)
info_df = merge(x = info_df, y = LLRscore_df[c("SAMPLE_ID","y_pred")], by = "SAMPLE_ID")
info_df = merge(x = info_df, y = RF6score_df[c("SAMPLE_ID","y_pred")], by = "SAMPLE_ID")
colnames(info_df) = c(renamed_cols, c("LLRscore","RF6score"))

CancerTypes = unique(info_df$CancerType)

```


# Pan-cancer K-M PFS curves of LLRscorehigh-TMBhigh VS LLRscorehigh-TMBlow VS...
```{r}
cutoffType = 'AbsoluteCutoff' # 'AbsoluteCutoff'  'PercentileCutoff80'  'PercentileCutoff50'
if (cutoffType == 'PercentileCutoff80'){
  TMB_Percentile = 0.8 
}else if (cutoffType == 'PercentileCutoff50'){
  TMB_Percentile = 0.5 
}


if (cutoffType == 'AbsoluteCutoff'){
  info_df = info_df %>% group_by(CancerType) %>% mutate(
         LLRscore_01 = ifelse(LLRscore >= absoluteCutoff, "high", "low"),
         RF6score_01 = ifelse(RF6score >= 0.27, "high", "low"),
         TMB_01 = ifelse(TMB >= 10, "high", "low"),
         )
}else{
  info_df = info_df %>% group_by(CancerType) %>% mutate(
         LLRscore_01 = ifelse(LLRscore >= quantile(LLRscore, 0.5), "high", "low"),
         RF6score_01 = ifelse(RF6score >= quantile(RF6score, 0.6), "high", "low"),
         TMB_01 = ifelse(TMB >= quantile(TMB, TMB_Percentile), "high", "low"),
         )
}

plot_df <- info_df %>%
  group_by(CancerType) %>% 
      mutate(LLR6_TMB_group = paste0(LLRscore_01, " LLRscore, ", TMB_01, " TMB"))


pdf(paste0(result_dir,"PanCancer_",cancer_type,'_',LLRmodelNA,"_TMB_PFScurve_",cutoffType,".pdf"), width = 5*0.8, height = 3.8, bg = "transparent", onefile = F) 
plot_font_size = 9
risktable_font_size = 3
ggsurvplot(survfit(Surv(PFS_Months, PFS_Event) ~ LLR6_TMB_group, data = plot_df), pval = F, 
          palette = plasma(4),
          risk.table=TRUE,
          risk.table.height = 0.25, # Useful to change when you have multiple groups
          risk.table.col="black",
          risk.table.y.text = FALSE, 
          tables.y.text = FALSE, 
          tables.theme = theme_cleantable(),
         xlab = 'Time (months)', ylab = 'PFS probability',
         legend = c(0.35, 0.88),
         legend.text = element_text(size = plot_font_size, color = "black"),
         legend.labs = c("LORIS-H/TMB-H", "LORIS-H/TMB-L", "LORIS-L/TMB-H", "LORIS-L/TMB-L"),
         ggtheme = theme(
                         panel.background = element_rect(fill='white'), #transparent panel bg
                         plot.background = element_rect(fill='white', color=NA), #transparent plot bg
                         legend.background = element_rect(fill='transparent', color=NA), #transparent legend bg
                         legend.box.background = element_rect(fill='transparent', color=NA), #transparent legend panel
                         legend.key = element_rect(fill='transparent', color=NA),
                         plot.margin = unit(c(0.2, 0.2, 0.2, 0.2),"cm"),
                         panel.grid.major = element_blank(), 
                         panel.grid.minor = element_blank(),
                         panel.border = element_blank(),
                         axis.line = element_line(colour = "black"),
                         axis.text.x = element_text(colour = "black", size = plot_font_size),
                        axis.text.y = element_text(colour = "black", size = plot_font_size),
                        axis.title.x = element_text(size = plot_font_size),
                        axis.title.y = element_text(size = plot_font_size),
                        legend.text = element_text(size = plot_font_size),
                        legend.title = element_text(size = plot_font_size),
                        plot.title = element_text(size = plot_font_size),
                        plot.subtitle = element_text(size = plot_font_size),
                        plot.caption = element_text(size = plot_font_size)
                         ),  # top, right, bot, left
        risk.table.fontsize = risktable_font_size,
        pval.method = T) +
guides(color=guide_legend(ncol =1)) +
labs(color = "")
dev.off()

##### HR and p-value comparison
cancerData=data.frame(info_df$LLRscore_01,info_df$PFS_Months,info_df$PFS_Event)
colnames(cancerData) = c("Score", "PFS_Months", "PFS_Event")
sfit = survfit(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
scox = coxph(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
HR_CI = exp(confint(scox))
print(paste(c('LLR6 PFS: ', format(P_value, scientific = TRUE),HR_value), collapse= " "))

cancerData=data.frame(info_df$TMB_01,info_df$PFS_Months,info_df$PFS_Event)
colnames(cancerData) = c("Score", "PFS_Months", "PFS_Event")
sfit = survfit(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
scox = coxph(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
HR_CI = exp(confint(scox))
print(paste(c('TMB PFS: ', P_value,HR_value), collapse= " "))

##### p values between 4 different groups
plot_df_temp = plot_df[plot_df$LLR6_TMB_group=='high LLRscore, low TMB' | plot_df$LLR6_TMB_group=='low LLRscore, low TMB', ]
scox = coxph(Surv(plot_df_temp$PFS_Months,plot_df_temp$PFS_Event)~LLR6_TMB_group, data=plot_df_temp)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
print(paste(c(format(P_value, scientific = TRUE),HR_value), collapse= " "))

```


# Pan-cancer K-M OS curves of LLRscorehigh-TMBhigh VS LLRscorehigh-TMBlow VS...
```{r}

cutoffType = 'PercentileCutoff80' # 'AbsoluteCutoff'  'PercentileCutoff80'  'PercentileCutoff50'
if (cutoffType == 'PercentileCutoff80'){
  TMB_Percentile = 0.8 
}else if (cutoffType == 'PercentileCutoff50'){
  TMB_Percentile = 0.5 
}


if (cutoffType == 'AbsoluteCutoff'){
  info_df = info_df %>% group_by(CancerType) %>% mutate(
         LLRscore_01 = ifelse(LLRscore >= absoluteCutoff, "high", "low"),
         RF6score_01 = ifelse(RF6score >= 0.27, "high", "low"),
         TMB_01 = ifelse(TMB >= 10, "high", "low"),
         )
}else{
  info_df = info_df %>% group_by(CancerType) %>% mutate(
         LLRscore_01 = ifelse(LLRscore >= quantile(LLRscore, 0.5), "high", "low"),
         RF6score_01 = ifelse(RF6score >= quantile(RF6score, 0.6), "high", "low"),
         TMB_01 = ifelse(TMB >= quantile(TMB, TMB_Percentile), "high", "low"),
         )
}


plot_df <- info_df %>%
  group_by(CancerType) %>% 
      mutate(LLR6_TMB_group = paste0(LLRscore_01, " LLRscore, ", TMB_01, " TMB"))


pdf(paste0(result_dir,"PanCancer_",cancer_type,'_',LLRmodelNA,"_TMB_OScurve_",cutoffType,".pdf"), width = 5*0.8, height = 3.8, onefile = F) 
plot_font_size = 9
risktable_font_size = 3
ggsurvplot(survfit(Surv(OS_Months, OS_Event) ~ LLR6_TMB_group, data = plot_df), pval = F, 
          palette = plasma(4),
          risk.table=TRUE,
          risk.table.height = 0.25, # Useful to change when you have multiple groups
          risk.table.col="black",
          risk.table.y.text = FALSE, 
          tables.y.text = FALSE, 
          tables.theme = theme_cleantable(),
         xlab = 'Time (months)', ylab = 'OS probability',
         legend = c(0.45, 0.88),
         legend.text = element_text(size = plot_font_size, color = "black"),
         legend.labs = c("LORIS-H/TMB-H", "LORIS-H/TMB-L", "LORIS-L/TMB-H", "LORIS-L/TMB-L"),
         ggtheme = theme(
                         panel.background = element_rect(fill='white'), #transparent panel bg
                         plot.background = element_rect(fill='white', color=NA), #transparent plot bg
                         legend.background = element_rect(fill='transparent', color=NA), #transparent legend bg
                         legend.box.background = element_rect(fill='transparent', color=NA), #transparent legend panel
                         legend.key = element_rect(fill='transparent', color=NA),
                         plot.margin = unit(c(0.2, 0.2, 0.2, 0.2),"cm"),
                         panel.grid.major = element_blank(), 
                         panel.grid.minor = element_blank(),
                         panel.border = element_blank(),
                         axis.line = element_line(colour = "black"),
                         axis.text.x = element_text(colour = "black", size = plot_font_size),
                        axis.text.y = element_text(colour = "black", size = plot_font_size),
                        axis.title.x = element_text(size = plot_font_size),
                        axis.title.y = element_text(size = plot_font_size),
                        legend.text = element_text(size = plot_font_size),
                        legend.title = element_text(size = plot_font_size),
                        plot.title = element_text(size = plot_font_size),
                        plot.subtitle = element_text(size = plot_font_size),
                        plot.caption = element_text(size = plot_font_size)
                         ),  # top, right, bot, left
        risk.table.fontsize = risktable_font_size,
        pval.method = T) +
guides(color=guide_legend(ncol =1)) +
labs(color = "")
dev.off()

##### HR and p-value comparison
cancerData=data.frame(info_df$LLRscore_01,info_df$OS_Months,info_df$OS_Event)
colnames(cancerData) = c("Score", "OS_Months", "OS_Event")
sfit = survfit(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
scox = coxph(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
HR_CI = exp(confint(scox))
print(paste(c('LLR6 OS: ', format(P_value, scientific = TRUE),HR_value), collapse= " "))

cancerData=data.frame(info_df$TMB_01,info_df$OS_Months,info_df$OS_Event)
colnames(cancerData) = c("Score", "OS_Months", "OS_Event")
sfit = survfit(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
scox = coxph(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
HR_CI = exp(confint(scox))
print(paste(c('TMB OS: ', P_value,HR_value), collapse= " "))

##### p values between 4 different groups
plot_df_temp = plot_df[plot_df$LLR6_TMB_group=='low LLRscore, high TMB' | plot_df$LLR6_TMB_group=='low LLRscore, low TMB', ]
scox = coxph(Surv(plot_df_temp$OS_Months,plot_df_temp$OS_Event)~LLR6_TMB_group, data=plot_df_temp)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
Z_value=scox_coef[4]
P_value=scox_coef[5]
print(paste(c(P_value,HR_value), collapse= " "))

```


# Binarize LLR6 and TMB values
```{r}
info_df = info_df %>% group_by(CancerType) %>% mutate(
         LLRscore_01 = ifelse(LLRscore >= absoluteCutoff, "high", "low"),
         RF6score_01 = ifelse(RF6score >= 0.27, "high", "low"),
         TMB_01 = ifelse(TMB >= 10, "high", "low"),
         LLRscore_01_2 = ifelse(LLRscore >= quantile(LLRscore, 0.5), "high", "low"),
         RF6score_01_2 = ifelse(RF6score >= quantile(RF6score, 0.6), "high", "low"),
         TMB_01_2 = ifelse(TMB >= quantile(TMB, 0.8), "high", "low"),
         )
info_df$LLRscore_01 <- relevel(factor(info_df$LLRscore_01), ref = "low")
info_df$RF6score_01 <- relevel(factor(info_df$RF6score_01), ref = "low")
info_df$TMB_01 <- relevel(factor(info_df$TMB_01), ref = "low")
info_df$LLRscore_01_2 <- relevel(factor(info_df$LLRscore_01_2), ref = "low")
info_df$RF6score_01_2 <- relevel(factor(info_df$RF6score_01_2), ref = "low")
info_df$TMB_01_2 <- relevel(factor(info_df$TMB_01_2), ref = "low")

```

# Univariate/Multivariable hazard ratio Forest plot (OS/PFS, continuous/binary)  HR_each_cancer
```{r}
survival_type = 'PFS' # 'PFS'  'OS'
HR_type = 'Multi' # Multi Uni
Var_type = 'binaryPercent' # continuous  binaryAbsolute binaryPercent
testVAR0 = 'LLRscore' # RF6score LLRscore TMB
if (Var_type == 'binaryAbsolute'){
  testVAR = paste0(testVAR0, '_01')
}else if (Var_type == 'binaryPercent'){
  testVAR = paste0(testVAR0, '_01_2')
}
testVAR_HR = testVAR
if (grepl("binary", Var_type)){
  testVAR_HR = paste0(testVAR,'high')
}
if (HR_type == 'Multi'){
  var_names = c(testVAR,'Drug', 'Age','Year_IO')
}else{
  var_names = c(testVAR)
}


HR_pancancer = data.frame()
cancerType = 'All'
data_for_HR = info_df
formula_obj <- as.formula(paste(paste0("Surv(",survival_type,"_Months, ",survival_type,"_Event) ~CancerType+"), paste(var_names, collapse = " + ")))
model = coxph(formula_obj, data = data_for_HR)

HR_ct = exp(model$coefficients[[testVAR_HR]])
HR_ct_025 = exp(confint(model))[testVAR_HR,1]
HR_ct_975 = exp(confint(model))[testVAR_HR,2]
p_ct = summary(model)$coefficients[testVAR_HR,5]

HR_pancancer = rbind(HR_pancancer,data.frame(HR_ct,HR_ct_025,HR_ct_975,p_ct,dim(data_for_HR)[1],cancerType))
colnames(HR_pancancer) = c(paste0('HR_',testVAR),paste0('HR_',testVAR,'025'),paste0('HR_',testVAR,'975'),paste0('p-val_',testVAR),'No. of patients', 'CancerType')

HR_each_cancer = data.frame()
formula_obj <- as.formula(paste(paste0("Surv(",survival_type,"_Months, ",survival_type,"_Event) ~"), paste(var_names, collapse = " + ")))
for (ct in CancerTypes){
  cancerType = ct
  model = coxph(formula_obj, data = data_for_HR[data_for_HR$CancerType==ct,])
  HR_ct = exp(model$coefficients[[testVAR_HR]])
  HR_ct_025 = exp(confint(model))[testVAR_HR,1]
  HR_ct_975 = exp(confint(model))[testVAR_HR,2]
  p_ct = summary(model)$coefficients[testVAR_HR,5]
  HR_each_cancer = rbind(HR_each_cancer,data.frame(HR_ct,HR_ct_025,HR_ct_975,p_ct,dim(data_for_HR[data_for_HR$CancerType==ct,])[1],cancerType))
}
colnames(HR_each_cancer) = c(paste0('HR_',testVAR),paste0('HR_',testVAR,'025'),paste0('HR_',testVAR,'975'),paste0('p-val_',testVAR),'No. of patients', 'CancerType')


command_str <- paste0("HR_each_cancer = arrange(HR_each_cancer, desc(HR_",testVAR,"))") # desc(is.na(HR_",testVAR,")),
eval(parse(text = command_str))
cancerOrder = HR_each_cancer$CancerType

HR_each_cancer = rbind(HR_each_cancer,HR_pancancer)



plot_data <- tibble::tibble(mean  = HR_each_cancer[[paste0('HR_',testVAR)]],
                            lower = HR_each_cancer[[paste0('HR_',testVAR,'025')]],
                            upper = HR_each_cancer[[paste0('HR_',testVAR,'975')]],
                            cancer = HR_each_cancer$CancerType,
                            pval = HR_each_cancer[[paste0('p-val_',testVAR)]],
                            OR = round(HR_each_cancer[[paste0('HR_',testVAR)]], 2),
                            size = HR_each_cancer$`No. of patients`)

options(scipen = 999)
P_value_raw = c(plot_data[['pval']], HR_pancancer[[paste0('p-val_',testVAR)]])


pval_vec = vector("character", length(P_value_raw))
for (i in 1:length(P_value_raw)){
  pval = P_value_raw[i]
  if (pval>=0.1){
    pval_vec[i] = as.character(round(pval,2))
  }else{
    pval_vec[i] = format(pval, scientific = TRUE, digits = 2)  # Format for scientific notation
  }
}

plot_data$pval = pval_vec[1:length(pval_vec)-1]

if (testVAR=='LLRscore'){ # testVAR=='LLRscore'
  xmin_lim = 0
  xmax_lim = 10
  breaks_x = c(0,1,10)
}else if (testVAR=="LLRscore_01" | testVAR=="LLRscore_01_2"){ # testVAR=='LLRscore'
  xmin_lim = 0
  xmax_lim = 4
  breaks_x = c(0,1,2,3,4)
}else if (testVAR=='TMB'){ # testVAR=='TMB'
  xmin_lim = 0
  xmax_lim = 2
  breaks_x = c(0,1,2)
}else if (testVAR=="TMB_01" | testVAR=="TMB_01_2"){ # testVAR=='LLRscore'
  xmin_lim = 0
  xmax_lim = 4
  breaks_x = c(0,1,2,3,4)
}
labels_x = breaks_x


fontSize = 1.2

pdf_file <- paste0(result_dir,paste0("PanCancer_Forest_",survival_type,"_",HR_type,"_",Var_type,"_",testVAR0,".pdf"))
fig_width = 5
fig_height = 5.5
pdf(pdf_file, onefile=FALSE, width = fig_width,height=fig_height)


plot_data %>%
  forestplot(labeltext = c("cancer", "size", "pval"),
             graph.pos = 3,
             boxsize = 0.35,
             #line.margin = .30, # We need to add this to avoid crowding
             vertices = TRUE,
             clip = c(xmin_lim, xmax_lim),
             xlog = FALSE,
             zero = 1,
             #txt_gp = fpTxtGp(cex=0.75),
             txt_gp = fpTxtGp(ticks=gpar(cex=fontSize),xlab=gpar(cex=fontSize),label=gpar(cex=fontSize),legend=gpar(cex=fontSize),title=gpar(cex=fontSize)), # label, summary, xlab, title, ticks legend
             xlab = "Hazard ratio",
             xticks = breaks_x,
             #lineheight=unit(1,'cm'),
             graphwidth = unit(3, "cm"),
             lineheight = unit(2, "cm")
             ) %>%
  fp_set_style(box = "black",
               line = "black",
               summary = "black") %>%
  fp_add_header(cancer = c("Cancer type")|> fp_txt_plain() |> fp_align_center(),
                size = c("Samples")|> fp_txt_plain() |> fp_align_center(), # "No. of patients"
                pval = c("P-value")|> fp_txt_plain() |> fp_align_center())
#par(mar=c(10, 10, 10, 10))

dev.off()

```

# X-year PFS comparison for score-low VS score-high patients (X = 0.5,1,2,3,4,5 years)
```{r}
testVAR = 'TMB_01_2' # RF6score_01 LLRscore_01 TMB_01 RF6score_01_2 LLRscore_01_2 TMB_01_2

if (grepl('_01_2', testVAR)){
  testVAR_NA = paste0(substr(testVAR, 1, nchar(testVAR) - 5),'_PercentileCutoff')
}else{
  testVAR_NA = paste0(substr(testVAR, 1, nchar(testVAR) - 3),'_AbsoluteCutoff')
}

PFSrate_score = data.frame()

for (PFS_time in c(6,12,24,36,48,60)){
  month_PFS = PFS_time
  for (ct in CancerTypes){
    
    data_ct = info_df[info_df['CancerType']==ct,]
    cancerData=data.frame(data_ct[testVAR],data_ct$PFS_Months,data_ct$PFS_Event)
    colnames(cancerData) = c("Score", "PFS_Months", "PFS_Event")
    sfit = survfit(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
    PFS_rates = summary(sfit,time=c(month_PFS),extend=TRUE)$surv
    PFSrate_score = rbind(PFSrate_score,data.frame(PFS_rates[1],PFS_rates[2],month_PFS,row.names = ct))
  }
  cancerData=data.frame(info_df[testVAR],info_df$PFS_Months,info_df$PFS_Event)
  colnames(cancerData) = c("Score", "PFS_Months", "PFS_Event")
  sfit = survfit(Surv(PFS_Months,PFS_Event) ~ Score, data=cancerData)
  PFS_rates = summary(sfit,time=c(month_PFS),extend=TRUE)$surv
  PFSrate_score = rbind(PFSrate_score, data.frame(PFS_rates[1],PFS_rates[2],month_PFS,row.names = 'All'))
}
colnames(PFSrate_score) = c('PFSR_L','PFSR_H','Time')

PFSrate_score['CANCER_TYPE'] = rownames(PFSrate_score)




################ plot legend
PFSrate_score_6month = PFSrate_score[PFSrate_score$Time==6,]
ggpaired(PFSrate_score_6month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Half year", legend.title = "Cancer type", ylab = "PFS probability") + # palette = my.cols
  theme(legend.position = "right", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right angle = 30
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("Xyear_Survival_legend.pdf")), width = 5, height = 5) # width = 2.3, height = 2.7

################ 6-month PFSR paired plot
PFSrate_score_6month = PFSrate_score[PFSrate_score$Time==6,]
ggpaired(PFSrate_score_6month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Half year", legend.title = "Cancer type", ylab = "PFS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right angle = 30
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_6month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 6-month PFS rate:",median(PFSrate_score_6month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_6month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_6month$PFSR_H, PFSrate_score_6month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('6-month Wilcoxon p-value: ',stat_result$p.value))

################ 12-month PFSR paired plot
PFSrate_score_12month = PFSrate_score[PFSrate_score$Time==12,]
ggpaired(PFSrate_score_12month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "One year", legend.title = "Cancer type", ylab = "PFS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_12month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 12-month PFS rate:",median(PFSrate_score_12month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_12month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_12month$PFSR_H, PFSrate_score_12month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('12-month Wilcoxon p-value: ',stat_result$p.value))

################ 24-month PFSR paired plot
PFSrate_score_24month = PFSrate_score[PFSrate_score$Time==24,]
ggpaired(PFSrate_score_24month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Two years", legend.title = "Cancer type", ylab = "PFS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_24month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 24-month PFS rate:",median(PFSrate_score_24month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_24month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_24month$PFSR_H, PFSrate_score_24month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('24-month Wilcoxon p-value: ',stat_result$p.value))

################ 36-month PFSR paired plot
PFSrate_score_36month = PFSrate_score[PFSrate_score$Time==36,]
ggpaired(PFSrate_score_36month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Three years", legend.title = "Cancer type", ylab = "PFS probability") +
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_36month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 36-month PFS rate:",median(PFSrate_score_36month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_36month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_36month$PFSR_H, PFSrate_score_36month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('36-month Wilcoxon p-value: ',stat_result$p.value))

################ 48-month PFSR paired plot
PFSrate_score_48month = PFSrate_score[PFSrate_score$Time==48,]
ggpaired(PFSrate_score_48month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Four years", legend.title = "Cancer type", ylab = "PFS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_48month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 48-month PFS rate:",median(PFSrate_score_48month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_48month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_48month$PFSR_H, PFSrate_score_48month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('48-month Wilcoxon p-value: ',stat_result$p.value))

################ 60-month PFSR paired plot
PFSrate_score_60month = PFSrate_score[PFSrate_score$Time==60,]
ggpaired(PFSrate_score_60month %>% rename(`High` = PFSR_H,
                                `Low` = PFSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Five years", legend.title = "Cancer type", ylab = "PFS probability") +
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("PFS_60month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 60-month PFS rate:",median(PFSrate_score_60month$PFSR_H, na.rm = TRUE) - median(PFSrate_score_60month$PFSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(PFSrate_score_60month$PFSR_H, PFSrate_score_60month$PFSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('60-month Wilcoxon p-value: ',stat_result$p.value))

```

# X-year OS comparison for score-low VS score-high patients (X = 0.5,1,2,3,4,5 years)
```{r}
testVAR = 'TMB_01_2' # RF6score_01 LLRscore_01 TMB_01 RF6score_01_2 LLRscore_01_2 TMB_01_2

if (grepl('_01_2', testVAR)){
  testVAR_NA = paste0(substr(testVAR, 1, nchar(testVAR) - 5),'_PercentileCutoff')
}else{
  testVAR_NA = paste0(substr(testVAR, 1, nchar(testVAR) - 3),'_AbsoluteCutoff')
}

OSrate_score = data.frame()

for (os_time in c(6,12,24,36,48,60)){
  month_OS = os_time
  for (ct in CancerTypes){
    
    data_ct = info_df[info_df['CancerType']==ct,]
    cancerData=data.frame(data_ct[testVAR],data_ct$OS_Months,data_ct$OS_Event)
    colnames(cancerData) = c("Score", "OS_Months", "OS_Event")
    sfit = survfit(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
    OS_rates = summary(sfit,time=c(month_OS),extend=TRUE)$surv
    OSrate_score = rbind(OSrate_score,data.frame(OS_rates[1],OS_rates[2],month_OS,row.names = ct))
  }
  cancerData=data.frame(info_df[testVAR],info_df$OS_Months,info_df$OS_Event)
  colnames(cancerData) = c("Score", "OS_Months", "OS_Event")
  sfit = survfit(Surv(OS_Months,OS_Event) ~ Score, data=cancerData)
  OS_rates = summary(sfit,time=c(month_OS),extend=TRUE)$surv
  OSrate_score = rbind(OSrate_score, data.frame(OS_rates[1],OS_rates[2],month_OS,row.names = 'All'))
}
colnames(OSrate_score) = c('OSR_L','OSR_H','Time')

OSrate_score['CANCER_TYPE'] = rownames(OSrate_score)


################ 6-month OSR paired plot
OSrate_score_6month = OSrate_score[OSrate_score$Time==6,]
ggpaired(OSrate_score_6month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Half year", legend.title = "Cancer type", ylab = "OS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right angle = 30
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_6month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 6-month OS rate:",median(OSrate_score_6month$OSR_H, na.rm = TRUE) - median(OSrate_score_6month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_6month$OSR_H, OSrate_score_6month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('6-month Wilcoxon p-value: ',stat_result$p.value))

################ 12-month OSR paired plot
OSrate_score_12month = OSrate_score[OSrate_score$Time==12,]
ggpaired(OSrate_score_12month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "One year", legend.title = "Cancer type", ylab = "OS probability") +
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_12month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 12-month OS rate:",median(OSrate_score_12month$OSR_H, na.rm = TRUE) - median(OSrate_score_12month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_12month$OSR_H, OSrate_score_12month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('12-month Wilcoxon p-value: ',stat_result$p.value))

################ 24-month OSR paired plot
OSrate_score_24month = OSrate_score[OSrate_score$Time==24,]
ggpaired(OSrate_score_24month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Two years", legend.title = "Cancer type", ylab = "OS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_24month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 24-month OS rate:",median(OSrate_score_24month$OSR_H, na.rm = TRUE) - median(OSrate_score_24month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_24month$OSR_H, OSrate_score_24month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('24-month Wilcoxon p-value: ',stat_result$p.value))

################ 36-month OSR paired plot
OSrate_score_36month = OSrate_score[OSrate_score$Time==36,]
ggpaired(OSrate_score_36month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Three years", legend.title = "Cancer type", ylab = "OS probability") +
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_36month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 36-month OS rate:",median(OSrate_score_36month$OSR_H, na.rm = TRUE) - median(OSrate_score_36month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_36month$OSR_H, OSrate_score_36month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('36-month Wilcoxon p-value: ',stat_result$p.value))

################ 48-month OSR paired plot
OSrate_score_48month = OSrate_score[OSrate_score$Time==48,]
ggpaired(OSrate_score_48month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Four years", legend.title = "Cancer type", ylab = "OS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_48month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 48-month OS rate:",median(OSrate_score_48month$OSR_H, na.rm = TRUE) - median(OSrate_score_48month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_48month$OSR_H, OSrate_score_48month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('48-month Wilcoxon p-value: ',stat_result$p.value))

################ 60-month OSR paired plot
OSrate_score_60month = OSrate_score[OSrate_score$Time==60,]
ggpaired(OSrate_score_60month %>% rename(`High` = OSR_H,
                                `Low` = OSR_L), 
         cond1 = "High", cond2 = "Low", line.color = "CANCER_TYPE", line.size = 1,
         fill = "gray65",
         palette = plasma(17), xlab = "Five years", legend.title = "Cancer type", ylab = "OS probability") + 
  theme(legend.position = "none", axis.text.x = element_text(angle = 0, hjust = 0.5)) + # right
  ylim(0, 1) 
ggsave(paste0(result_dir,paste0("OS_60month_Comparison_",testVAR_NA,".pdf")), width = 1.8*1.2, height = 2.4*0.9) # width = 2.3, height = 2.7
print(paste("Delta 60-month OS rate:",median(OSrate_score_60month$OSR_H, na.rm = TRUE) - median(OSrate_score_60month$OSR_L, na.rm = TRUE),sep=" "))
stat_result = wilcox.test(OSrate_score_60month$OSR_H, OSrate_score_60month$OSR_L, paired = TRUE, alternative = "two.sided", na.rm = TRUE)
print(paste0('60-month Wilcoxon p-value: ',stat_result$p.value))

```

# 6 groups of K-M PFS curves of LORIS / TMB (0-10%,10-20%,20-50%,50-80%,80-90%,90-100%-quantile)
```{r}
####### load data
dataType = 'PanCancer_all'  # PanCancer_nonNSCLC   PanCancer_all
LLRmodelNA = 'LLR6' #  'LLR6'   'TMB'
scalerType = 'StandardScaler'
if (LLRmodelNA == 'TMB'){
  scalerType = 'None'
}

test1_Score_df = read_excel(paste0(result_dir,dataType,'_',LLRmodelNA,'_Scaler(',scalerType,')_prediction.xlsx'), "1", col_names = TRUE)
test2_Score_df = read_excel(paste0(result_dir,dataType,'_',LLRmodelNA,'_Scaler(',scalerType,')_prediction.xlsx'), "2")
Score_df = rbind(test1_Score_df,test2_Score_df)

selected_cols = c("SAMPLE_ID","TMB","Systemic_therapy_history","CancerType","Age","Drug","Sex","PFS_Event","PFS_Months","OS_Event","OS_Months")

test1_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "Chowell_test")
test1_info_df = test1_info_df[selected_cols]
test2_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK1")
test2_info_df = test2_info_df[selected_cols]
info_df = rbind(test1_info_df,test2_info_df)
info_df = merge(x = info_df, y = Score_df[c("SAMPLE_ID","y_pred")], by = "SAMPLE_ID")

colnames(info_df) = c(selected_cols, c("Score"))

quantile_plot = info_df %>% group_by(CancerType) %>%
  mutate(Score_perct = ntile(Score, n = n()) / n()) # ntile(Score, length(breaks)-1, breaks=breaks)

quantile_plot = quantile_plot %>% 
  mutate(Score_01 = ifelse(Score_perct > 0.9, 6, ifelse(Score_perct > 0.8, 5, ifelse(Score_perct > 0.5, 4, ifelse(Score_perct > 0.2, 3, ifelse(Score_perct > 0.1, 2, ifelse(Score_perct >= 0, 1, 0))))))) # ntile(Score, length(breaks)-1, breaks=breaks)

fontSize = 13
survp=ggsurvplot(
  survfit(Surv(PFS_Months, PFS_Event) ~ Score_01, data = quantile_plot),
  data = quantile_plot,
  size = 1,                 # change line size
  palette = plasma(6),
  conf.int = FALSE,          # Add confidence interval
  pval = FALSE,              # Add p-value
  ylim=c(0,1),xlim=c(-1,60),
  xlab = "Time (months)", ylab = 'PFS probability',
  break.time.by = 20,
  risk.table=TRUE,
  risk.table.height = 0.3, # Useful to change when you have multiple groups
  risk.table.col="black",
  risk.table.y.text = FALSE,
  tables.y.text = FALSE, 
  tables.theme = theme_cleantable(),
  legend.labs = c("0-10%", "10-20%", "20-50%", "50-80%", "80-90%", "90-100%"),
  legend.title="",
  legend = c(0.65, 0.85), # legend relative position
  font.main = c(fontSize),
  font.caption = c(fontSize),
  font.legend = c(fontSize),
  font.tickslab = c(fontSize),
  font.x = c(fontSize),
  font.y = c(fontSize),
 ggtheme = theme(
                 panel.background = element_rect(fill='transparent'), #transparent panel bg
                 plot.background = element_rect(fill='transparent', color=NA), #transparent plot bg
                 legend.background = element_rect(fill='transparent', color=NA), #transparent legend bg
                 legend.box.background = element_rect(fill='transparent', color=NA), #transparent legend panel
                 legend.key = element_rect(fill='transparent', color=NA),
                 plot.margin = unit(c(0.2, 0.2, 0.2, 0.2),"cm"),
                 panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                 panel.border = element_blank(),axis.line = element_line(colour = "black"),
                 axis.text.x = element_text(colour="black"),axis.text.y = element_text(colour="black")
                 ),  # top, right, bot, left
)
pdf(paste0(result_dir,dataType,'_',LLRmodelNA,"_PFS_curves.pdf"),width = 5, height = 6)
print(survp, newpage = FALSE)
dev.off()



##### p values between 5 different groups
quantile_plot_temp = quantile_plot[quantile_plot$Score_01==6 | quantile_plot$Score_01==1, ]

quantile_plot_temp = quantile_plot_temp %>% mutate(
         Score_01 = ifelse(Score_01 == 6, "high", "low"),
         )
quantile_plot_temp$Score_01 <- relevel(factor(quantile_plot_temp$Score_01), ref = "high")

scox = coxph(Surv(quantile_plot_temp$PFS_Months,quantile_plot_temp$PFS_Event)~Score_01, data=quantile_plot_temp)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
HR_CI = exp(confint(scox))
Z_value=scox_coef[4]
P_value= pnorm(Z_value, lower.tail = FALSE) # to test if higher score has better survival
print(paste0('HR_value: ', round(HR_value,2), '.  P_value: ', P_value))
```

# 6 groups of K-M OS curves of LORIS / TMB (0-10%,10-20%,20-50%,50-80%,80-90%,90-100%-quantile)
```{r}
####### load data
dataType = 'PanCancer_nonNSCLC'  # PanCancer_nonNSCLC   PanCancer_all
LLRmodelNA = 'TMB' #  'LLR6'   'TMB'
scalerType = 'StandardScaler'
if (LLRmodelNA == 'TMB'){
  scalerType = 'None'
}

test1_Score_df = read_excel(paste0(result_dir,dataType,'_',LLRmodelNA,'_Scaler(',scalerType,')_prediction.xlsx'), "1", col_names = TRUE)
test2_Score_df = read_excel(paste0(result_dir,dataType,'_',LLRmodelNA,'_Scaler(',scalerType,')_prediction.xlsx'), "2")
Score_df = rbind(test1_Score_df,test2_Score_df)

selected_cols = c("SAMPLE_ID","TMB","Systemic_therapy_history","CancerType","Age","Drug","Sex","PFS_Event","PFS_Months","OS_Event","OS_Months")

test1_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "Chowell_test")
test1_info_df = test1_info_df[selected_cols]
test2_info_df = read_excel(paste0(input_dir,'AllData.xlsx'), "MSK1")
test2_info_df = test2_info_df[selected_cols]
info_df = rbind(test1_info_df,test2_info_df)
info_df = merge(x = info_df, y = Score_df[c("SAMPLE_ID","y_pred")], by = "SAMPLE_ID")

colnames(info_df) = c(selected_cols, c("Score"))

quantile_plot = info_df %>% group_by(CancerType) %>%
  mutate(Score_perct = ntile(Score, n = n()) / n()) # ntile(Score, length(breaks)-1, breaks=breaks)

quantile_plot = quantile_plot %>% 
  mutate(Score_01 = ifelse(Score_perct > 0.9, 6, ifelse(Score_perct > 0.8, 5, ifelse(Score_perct > 0.5, 4, ifelse(Score_perct > 0.2, 3, ifelse(Score_perct > 0.1, 2, ifelse(Score_perct >= 0, 1, 0))))))) # ntile(Score, length(breaks)-1, breaks=breaks)


fontSize = 13
survp=ggsurvplot(
  survfit(Surv(OS_Months, OS_Event) ~ Score_01, data = quantile_plot),
  data = quantile_plot,
  size = 1,                 # change line size
  palette = plasma(6),
  conf.int = FALSE,          # Add confidence interval
  pval = FALSE,              # Add p-value
  ylim=c(0,1),xlim=c(-1,60),
  xlab = "Time (months)", ylab = 'OS probability',
  break.time.by = 20,
  risk.table=TRUE,
  risk.table.height = 0.3, # Useful to change when you have multiple groups
  risk.table.col="black",
  risk.table.y.text = FALSE,
  tables.y.text = FALSE, 
  tables.theme = theme_cleantable(),
  legend.labs = c("0-10%", "10-20%", "20-50%", "50-80%", "80-90%", "90-100%"),
  legend.title="",
  legend = c(0.65, 0.85), # legend relative position
  font.main = c(fontSize),
  font.caption = c(fontSize),
  font.legend = c(fontSize),
  font.tickslab = c(fontSize),
  font.x = c(fontSize),
  font.y = c(fontSize),
 ggtheme = theme(
                 panel.background = element_rect(fill='transparent'), #transparent panel bg
                 plot.background = element_rect(fill='transparent', color=NA), #transparent plot bg
                 legend.background = element_rect(fill='transparent', color=NA), #transparent legend bg
                 legend.box.background = element_rect(fill='transparent', color=NA), #transparent legend panel
                 legend.key = element_rect(fill='transparent', color=NA),
                 plot.margin = unit(c(0.2, 0.2, 0.2, 0.2),"cm"),
                 panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
                 panel.border = element_blank(),axis.line = element_line(colour = "black"),
                 axis.text.x = element_text(colour="black"),axis.text.y = element_text(colour="black")
                 ),  # top, right, bot, left
)
pdf(paste0(result_dir,dataType,'_',LLRmodelNA,"_OS_curves.pdf"),width = 5, height = 6)
print(survp, newpage = FALSE)
dev.off()



##### p values between 5 different groups
quantile_plot_temp = quantile_plot[quantile_plot$Score_01==6 | quantile_plot$Score_01==1, ]

quantile_plot_temp = quantile_plot_temp %>% mutate(
         Score_01 = ifelse(Score_01 == 6, "high", "low"),
         )
quantile_plot_temp$Score_01 <- relevel(factor(quantile_plot_temp$Score_01), ref = "high")

scox = coxph(Surv(quantile_plot_temp$OS_Months,quantile_plot_temp$OS_Event)~Score_01, data=quantile_plot_temp)
scox_coef = summary(scox)$coefficients
HR_value = scox_coef[2] # hazard ratio
HR_CI = exp(confint(scox))
Z_value=scox_coef[4]
P_value= pnorm(Z_value, lower.tail = FALSE) # to test if higher score has better survival
print(paste0('HR_value: ', round(HR_value,2), '.  P_value: ', format(P_value, scientific = TRUE)))
```
